Update base POM to version 5. Merge of FC branch into master.

+++ CCSDS_MAL_TRANSPORT_FILE/src/main/java/esa/mo/mal/encoder/line/LineDecoder.java
@@ -293,7 +293,7 @@ public class LineDecoder implements MALDecoder
   {
     try
     {
-      return new Duration(Integer.parseInt(strVal));
+      return new Duration(Double.parseDouble(strVal));
     }
     catch (NumberFormatException ex)
     {
